+++ src/Telegram.Bot/Exceptions/BadRequestException.cs
@@ -3,30 +3,60 @@ using Telegram.Bot.Types;
 
 namespace Telegram.Bot.Exceptions
 {
+    /// <summary>
+    /// A base class for "Error 400: Bad request" API responses
+    /// </summary>
     public abstract class BadRequestException : ApiRequestException
     {
+        /// <inheritdoc />
         public override int ErrorCode => BadRequestErrorCode;
 
+        /// <summary>
+        /// Represent error code number
+        /// </summary>
         public const int BadRequestErrorCode = 400;
 
+        /// <summary>
+        /// Represent error description
+        /// </summary>
         public const string BadRequestErrorDescription = "Bad Request: ";
 
-        public BadRequestException(string message)
+        /// <summary>
+        /// Initializes a new instance of the <see cref="BadRequestException"/> class
+        /// </summary>
+        /// <param name="message">The message</param>
+        protected BadRequestException(string message)
             : base(message, BadRequestErrorCode)
         {
         }
 
-        public BadRequestException(string message, Exception innerException)
+        /// <summary>
+        /// Initializes a new instance of the <see cref="BadRequestException"/> class
+        /// </summary>
+        /// <param name="message">The message</param>
+        /// <param name="innerException">The inner exception</param>
+        protected BadRequestException(string message, Exception innerException)
             : base(message, BadRequestErrorCode, innerException)
         {
         }
 
-        public BadRequestException(string message, ResponseParameters parameters)
+        /// <summary>
+        /// Initializes a new instance of the <see cref="BadRequestException"/> class
+        /// </summary>
+        /// <param name="message">The message</param>
+        /// <param name="parameters">Response parameters</param>
+        protected BadRequestException(string message, ResponseParameters parameters)
             : base(message, BadRequestErrorCode, parameters)
         {
         }
 
-        public BadRequestException(string message, ResponseParameters parameters, Exception innerException)
+        /// <summary>
+        /// Initializes a new instance of the <see cref="BadRequestException"/> class
+        /// </summary>
+        /// <param name="message">The message</param>
+        /// <param name="parameters">Response parameters</param>
+        /// <param name="innerException">The inner exception</param>
+        protected BadRequestException(string message, ResponseParameters parameters, Exception innerException)
             : base(message, BadRequestErrorCode, parameters, innerException)
         {
         }