+++ test/Telegram.Bot.Tests.Integ/Inline Mode/InlineQueryTests.cs
@@ -28,7 +28,6 @@ namespace Telegram.Bot.Tests.Integ.Inline_Mode
         [ExecutionOrder(1)]
         public async Task Should_Answer_Inline_Query_With_Article()
         {
-            // ToDo: add exception: Bad Request: QUERY_ID_INVALID
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldAnswerInlineQueryWithArticle,
                 startInlineQuery: true);
 
@@ -749,7 +748,7 @@ namespace Telegram.Bot.Tests.Integ.Inline_Mode
 
             while (
                 !cancellationToken.IsCancellationRequested &&
-                (messageUpdate is default || chosenResultUpdate is default)
+                (messageUpdate == null || chosenResultUpdate == null)
             )
             {
                 await Task.Delay(1_000, cancellationToken);