+++ CHANGELOG.md
@@ -11,6 +11,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Exception `InvalidQueryIdException`
 
+### Changed
+
+- Access modifier of abstract class `BadRequestException` and `ForbiddenException` ctors to `protected`
+
+### Removed
+
+- Exception `BotBlockedException`
+- Exception `BotRestrictedException`
+- Exception `MissingParameterException`
+- Exception `NotEnoughRightsException`
+- Exception `WrongChatTypeException`
 
 ## [14.2.0-rc-452] - 2018-02-24